demos/02-agent-service/01-agent-basics/agent-basics-py/pyproject.toml
[project]
name = "diagram-agent-py"
version = "0.1.0"
description = "AI Agent demo with diagram generation"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "python-dotenv>=1.0.0",
    "azure-identity>=1.16.0",
    "azure-ai-projects>=2.0.0b2",
    "azure-ai-agents>=1.2.0b6",
    "pathlib>=1.0.1",
    "qrcode[pil]>=7.0",
    "azure-storage-blob>=12.0.0",
    "Pillow>=10.0.0"
]